backend/requirements.txt
# --- Web framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database ---
sqlalchemy==2.0.35

# --- Data / config ---
pydantic==2.9.2
pydantic-settings==2.5.2
pyyaml==6.0.2

# --- Auth (заготовка под JWT, понадобится позже) ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- WebSocket support (включён в uvicorn[standard], фиксируем явно) ---
websockets==13.1

# --- Dev / тесты ---
pytest==8.3.3
httpx==0.27.2

# --- CV-стек (шаг 4) ---
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
# YOLOv8 через ultralytics. Torch устанавливается отдельно (см. README):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
ultralytics==8.3.0
aiofiles==24.1.0
